fix pasting images not attaching them to the mail
squire can handle inserting inline images with blob urls just fine,
but the way it was done circumvented our own inline image handling.
